Add files via upload
High employee attrition rates can be detrimental to a company's productivity, morale, and financial stability. To address this, exploratory data analysis is conducted to identify patterns and trends in employee data, uncover significant factors leading to attrition, and enable the company to implement measures to retain employees and decrease attrition rates. Furthermore, predictive model using machine learning techniques in python is built, which can forecast whether an employee will attrite or not.